L12/agent ai claude/requirements.txt
# Gmail API
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Excel handling
openpyxl==3.1.2

# MCP Server
mcp>=1.0.0

# Utilities
python-dotenv==1.0.0
